+++ Framework/PWMapKit.framework/Headers/PWPointOfInterest.h
@@ -16,7 +16,7 @@
 @interface PWPointOfInterest : NSObject <PWMapPoint>
 
 /**
- *  A description of the point-of-interest.
+ *  A description of the point-of-interest. May be NULL.
  */
 @property (nonatomic, readonly) NSString *summary;
 
@@ -26,22 +26,22 @@
 @property (nonatomic, readonly, weak, nullable) PWFloor *floor;
 
 /**
- *  The `PWPointOfInterestType` object that the point-of-interest is a member of.
+ *  The `PWPointOfInterestType` object that the point-of-interest is a member of. Should not be NULL.
  */
 @property (nonatomic, readonly) PWPointOfInterestType *pointOfInterestType;
 
 /**
- *  The representative UIImage of the point-of-interest.
+ *  The representative UIImage of the point-of-interest. Will be NULL if image has not yet been loaded.
  */
 @property (nonatomic, readonly) UIImage *image;
 
 /**
- *  The URL of the point-of-interest image.
+ *  The URL of the point-of-interest image. May be NULL if using custom POI image loading.
  */
 @property (nonatomic, readonly) NSURL *imageURL;
 
 /**
- * Metadata associated with the point-of-interest.
+ * Metadata associated with the point-of-interest. May be NULL.
  */
 @property (nonatomic, readonly) NSDictionary *metaData;
 
@@ -72,12 +72,12 @@
 @property (nonatomic, readonly) NSInteger maxZoomLevel;
 
 /**
- * The title of the point-of-interest.
+ * The title of the point-of-interest. May be NULL.
  */
 @property (nonatomic, copy, readwrite) NSString *title;
 
 /**
- * The subtitle of the point-of-interest.
+ * The subtitle of the point-of-interest. May be NULL.
  */
 @property (nonatomic, copy, readwrite) NSString *subtitle;